requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
google-cloud-speech==2.21.0
python-dotenv==1.0.0
python-multipart==0.0.6